cantina_os/requirements.txt
# Core Dependencies
pydantic>=2.5.2  # Data validation and settings management
pyee>=11.0.1    # Event emitter for Python
python-dotenv>=1.0.0  # Environment variable management
aiohttp>=3.9.1  # Async HTTP client for API requests
httpx>=0.25.0   # Modern async HTTP client for API requests

# Audio Processing
sounddevice>=0.4.6  # Audio capture and playback
numpy>=1.24.0    # Audio data manipulation
deepgram-sdk>=2.12.0  # Streaming ASR
openai-whisper>=20231117  # Local ASR (optional)
python-vlc>=3.0.20000  # VLC media player integration for music playback

# LLM Integration
openai>=1.3.7  # GPT-4 integration
anthropic>=0.7.7  # Claude integration (optional)

# Speech Synthesis
elevenlabs>=0.2.26  # Text-to-speech
soundfile>=0.12.1  # Audio file reading/writing for sounddevice

# Hardware Control
pyserial>=3.5  # Serial communication for Arduino/LED control

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
aioresponses>=0.7.4  # For mocking async HTTP responses in tests
psutil>=5.9.0  # For resource monitoring in tests

# Development Tools
black>=23.11.0  # Code formatting
isort>=5.12.0   # Import sorting
mypy>=1.7.1     # Static type checking
ruff>=0.1.6     # Fast Python linter 